# web/openapi.json
{
  "components": {
    "schemas": {
      "AliasOut": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "alias",
          "language"
        ],
        "title": "AliasOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngredientOut": {
        "properties": {
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/AliasOut"
            },
            "title": "Aliases",
            "type": "array"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "density_g_per_ml": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Density G Per Ml"
          },
          "dietary_flags": {
            "items": {
              "type": "string"
            },
            "title": "Dietary Flags",
            "type": "array"
          },
          "gram_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Gram Weights",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preferred_measure": {
            "$ref": "#/components/schemas/PreferredMeasure"
          },
          "status": {
            "$ref": "#/components/schemas/IngredientStatus"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "preferred_measure",
          "status",
          "dietary_flags",
          "density_g_per_ml",
          "gram_weights",
          "aliases"
        ],
        "title": "IngredientOut",
        "type": "object"
      },
      "IngredientPatch": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "density_g_per_ml": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Density G Per Ml"
          },
          "dietary_flags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dietary Flags"
          },
          "gram_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gram Weights"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "preferred_measure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreferredMeasure"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngredientStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "IngredientPatch",
        "type": "object"
      },
      "IngredientStatus": {
        "enum": [
          "seeded",
          "unreviewed",
          "reviewed"
        ],
        "title": "IngredientStatus",
        "type": "string"
      },
...
      "MergeIn": {
        "properties": {
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "target_id"
        ],
        "title": "MergeIn",
        "type": "object"
      },
      "PreferredMeasure": {
        "enum": [
          "mass",
          "volume"
        ],
        "title": "PreferredMeasure",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Recipe Normalizer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/catalog/ingredients": {
      "get": {
        "operationId": "list_ingredients_api_catalog_ingredients_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IngredientStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientOut"
                  },
                  "title": "Response List Ingredients Api Catalog Ingredients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ingredients",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/ingredients/{ingredient_id}": {
      "get": {
        "operationId": "get_ingredient_api_catalog_ingredients__ingredient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingredient",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "operationId": "patch_ingredient_api_catalog_ingredients__ingredient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Ingredient",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/ingredients/{ingredient_id}/merge": {
      "post": {
        "operationId": "merge_ingredient_api_catalog_ingredients__ingredient_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Ingredient",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}